Update README.md
This is my personal Website to showcase my sample projects and resume; I build light-weight Websites and system intergrations for businesses such as free hosting, Content Management System, and custom POS plugins. This Repo is for hosting the Website at github pages; the Repo for code is in BitBucket. Need to merge the code changes made here back to BitBucket using pull request.